report_debug_agent/requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-chroma
langgraph
langchain-community
pypdf
python-dotenv
langgraph-checkpoint-sqlite
docx2txt
fastapi
uvicorn
pydantic
python-multipart
aiosqlite
langchain-groq
